files_for_help/files/mnt/user-data/outputs/bloomberg-markets-summarizer/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Video/Audio processing
yt-dlp>=2023.12.0
pydub>=0.25.1

# Transcription
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# AI/LLM
anthropic>=0.18.0
openai>=1.12.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging
colorlog>=6.8.0

# Optional: Email notifications
# sendgrid>=6.11.0
